# Main_service/docs/swagger.yaml
basePath: /api/v1
definitions:
  model.Chapter:
    properties:
      course_id:
        type: integer
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      order:
        type: integer
      updated_at:
        type: string
    type: object
  model.Course:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
  model.Lesson:
    properties:
      chapter_id:
        type: integer
      content:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      order:
        type: integer
      updated_at:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API Server for LMS Main Service
  title: LMS Main Service API
  version: "1.0"
paths:
  /chapters:
    post:
      consumes:
      - application/json
      description: Create a new chapter
      parameters:
      - description: Chapter data
        in: body
        name: chapter
        required: true
        schema:
          $ref: '#/definitions/model.Chapter'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Chapter'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create a new chapter
      tags:
      - chapters
  /chapters/{chapter_id}/lessons:
    get:
      description: Get all lessons associated with a specific chapter ID
      parameters:
      - description: Chapter ID
        in: path
        name: chapter_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Lesson'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get lessons by Chapter ID
      tags:
      - lessons
  /courses:
    get:
      description: Get all courses
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Course'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get all courses
      tags:
      - courses
    post:
      consumes:
      - application/json
      description: Create a new course
      parameters:
      - description: Course data
        in: body
        name: course
        required: true
        schema:
          $ref: '#/definitions/model.Course'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Course'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create a new course
      tags:
      - courses
  /courses/{id}:
    delete:
      description: Delete a course by ID
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Delete a course
      tags:
      - courses
    get:
      description: Get course by ID
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Course'
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get course by ID
      tags:
      - courses
    put:
      consumes:
      - application/json
      description: Update a course by ID
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: integer
      - description: Course data
        in: body
        name: course
        required: true
        schema:
          $ref: '#/definitions/model.Course'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Course'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Update a course
      tags:
      - courses
  /lessons:
    post:
      consumes:
      - application/json
      description: Create a new lesson
      parameters:
      - description: Lesson data
        in: body
        name: lesson
        required: true
        schema:
          $ref: '#/definitions/model.Lesson'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Lesson'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create a new lesson
      tags:
      - lessons
  /lessons/{id}:
    delete:
      description: Delete a lesson by ID
      parameters:
      - description: Lesson ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Delete a lesson
      tags:
      - lessons
    get:
      description: Get lesson by ID
      parameters:
      - description: Lesson ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Lesson'
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get lesson by ID
      tags:
      - lessons
    put:
      consumes:
      - application/json
      description: Update a lesson by ID
      parameters:
      - description: Lesson ID
        in: path
        name: id
        required: true
        type: integer
      - description: Lesson data
        in: body
        name: lesson
        required: true
        schema:
          $ref: '#/definitions/model.Lesson'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Lesson'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Update a lesson
      tags:
      - lessons
swagger: "2.0"